# openapi.yaml
openapi: 3.0.3
info:
  title: Haushalt API
  version: 0.1.0
  description: |
    Der Vertrag zwischen der Web-App und dem Go-Dienst.

    Diese Datei ist die Wahrheit: Go-Typen und Server-Rümpfe werden daraus
    erzeugt (`make generate` in `api/`), der TypeScript-Client ebenso
    (`npm run generate` in `web/`). Wer ein Feld ändert, ändert es hier —
    alles andere folgt.

    Feldnamen sind deutsch und in Kleinschreibung mit Unterstrich, wie in
    der Vorlagen-Bibliothek. Die Domäne ist deutsch; eine halb übersetzte
    API wäre in beiden Sprachen falsch.

    `/healthz` steht bewusst nicht hier. Es gehört der Plattform, nicht dem
    Vertrag: Fly ruft es auf, kein Client.
servers:
  - url: https://haushalt-api.fly.dev
    description: Produktion
  - url: http://localhost:8080
    description: lokal
...
paths:
...
  /api/haushalte/{haushaltId}:
    patch:
      tags: [haushalt]
      summary: Einstellungen des Haushalts ändern
      description: |
        Nur planende Personen. Was hier steht, entscheidet, welche Aufgaben es
        im Haushalt überhaupt gibt — „kein Garten" nimmt das Rasenmähen für
        alle aus dem Plan.

        Gilt ab der nächsten Woche. Die laufende steht fest (ADR-0008) und
        wird nur auf Zuruf neu gerechnet.
      operationId: updateHaushalt
      parameters:
        - name: haushaltId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HaushaltAenderung"
      responses:
        "200":
          description: der Haushalt, wie er jetzt aussieht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Haushalt"
        "400":
          description: die Änderung ergibt keinen Haushalt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fehler"
        "403":
          description: das dürfen die planenden Personen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fehler"
        "404":
          description: diesen Haushalt gibt es nicht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fehler"
...
components:
  schemas:
...
    Fehler:
      type: object
      required: [fehler]
      properties:
        fehler:
          type: string
          description: für Menschen lesbar, deutsch
          example: die Woche "2026-W99" liegt außerhalb von 1..53
...
    HaushaltAenderung:
      type: object
      description: |
        Jedes Feld ist freiwillig. Was fehlt, bleibt stehen — der Unterschied
        zwischen „nicht mitgeschickt" und „geleert" ist der ganze Sinn einer
        Einstellungsseite.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        wohnform:
          type: string
          enum: [wohnung, haus]
        zimmer:
          type: integer
          minimum: 1
          maximum: 15
        baeder:
          type: integer
          minimum: 0
          maximum: 5
        garten:
          type: boolean
        auto:
          type: boolean
        haustiere:
          type: array
          maxItems: 6
          items:
            type: string
          description: Vollständige Liste. Leer heißt „keine mehr".
        zeitzone:
          type: string
...
    Haushalt:
      type: object
      required: [id, name, mitglieder]
      properties:
        id:
          type: string
          example: familie-a
        name:
          type: string
          example: Familie A
        mitglieder:
          type: array
          items:
            $ref: "#/components/schemas/Mitglied"
        meine_rolle:
          type: string
          enum: [planend, ausfuehrend, betreut]
          description: |
            Die Rolle des Aufrufers in diesem Haushalt. Fehlt bei den
            öffentlichen Beispielhaushalten und überall dort, wo er nicht
            Mitglied ist.
        ich:
          type: string
          description: |
            Kennung des Aufrufers als Person in diesem Haushalt. Damit weiß die
            Ansicht, welche Zeile seine eigene ist — etwa in den Einstellungen,
            wo jeder seine eigene Zeit setzen darf und sonst nichts.
        wohnform:
          type: string
          enum: [wohnung, haus]
        garten:
          type: boolean
        auto:
          type: boolean
        haustiere:
          type: array
          items:
            type: string
          description: |
            Garten, Auto und Haustiere entscheiden, welche Aufgaben es im
            Haushalt überhaupt gibt — ohne Garten kein Rasen, ohne Auto kein
            TÜV. Ohne diese Felder könnten die Einstellungen nicht zeigen, was
            angenommen wurde.
        zimmer:
          type: integer
        baeder:
          type: integer
    Mitglied:
      type: object
      required: [id, name, rolle]
      properties:
        id:
          type: string
          example: m-anna
        name:
          type: string
          example: Anna
        rolle:
          type: string
          enum: [planend, ausfuehrend, betreut]
          description: |
            `planend` plant und führt aus, `ausfuehrend` führt nur aus,
            `betreut` erzeugt Arbeit ohne welche zu übernehmen.
        hat_zugang:
          type: boolean
          description: |
            Ob diese Person sich anmelden kann. Steuert, wen man noch einladen
            kann — und macht sichtbar, wer bisher nur im Plan steht.
        geburtsjahr:
          type: integer
          description: |
            Fehlt bei Erwachsenen — dort ist „nicht gefragt" die richtige
            Antwort und nicht „unbekannt alt" (siehe Member.IsAdult im Planer).
        minuten:
          type: array
          minItems: 7
          maxItems: 7
          items:
            type: integer
          description: |
            Verfügbare Minuten je Wochentag, Index 0 = Montag. Beim Einrichten
            aus einer der drei Stufen geraten und hier korrigierbar. Die Zahlen
            tragen die gesamte Verteilung — deshalb stehen sie sichtbar da.
        betreuung:
          type: string
          enum: [keine, kita, schule]
          description: |
            Beim Einrichten aus dem Alter geraten und hier korrigierbar. An ihr
            hängt, welche Aufgaben der Haushalt überhaupt hat — ohne Kita-Kind
            keine Kita-Tasche.
...